Migrate consumers of LibrariesHandler, remove useless helper
